move PageHeading to /components/type/
