+++ frontend/src/metabase/components/type/PageHeading.info.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PageHeading from "metabase/components/PageHeading";
+import PageHeading from "metabase/components/type/PageHeading";
 
 export const component = PageHeading;
